+++ Assignment1/src/main/java/models/board.java
@@ -23,4 +23,8 @@ public class board {
         System.out.println("Column: " + col);
         System.out.println("Top Element: " + boardArray[col].peek());
     }*/
+
+    public void removeStack(int col){
+        boardArray[col].pop();
+    }
 }
